requirements.txt
# The app itself
-e .[tests]

django-formtools
psycopg2
